refactor(backend): improve validation in entities and update carbonSaved inside EcoActionType to match numeric(5,2) in DB

+++ backend/src/main/java/com/localzero/dto/EcoActionResponse.java
@@ -2,10 +2,12 @@ package com.localzero.dto;
 
 import com.localzero.model.enums.EcoActionCategory;
 
+import java.math.BigDecimal;
+
 public record EcoActionResponse(
         Long id,
         String date,
         String action,
         EcoActionCategory category,
-        Double carbonSaved
+        BigDecimal carbonSaved
 ) { }
